+++ js/components/dataset/card.vue
@@ -3,70 +3,52 @@
     <a v-if="dataset.organization" class="card-logo" :href="dataset.page">
         <img :alt="dataset.organization.name" :src="logo">
     </a>
+
     <img v-if="dataset.organization && dataset.organization.public_service"
         :src="certified" alt="certified" class="certified"
         v-popover="_('The identity of this public service is certified by %(certifier)s', certifier=config.SITE_AUTHOR)"
         :popover-title="_('Certified public service')"
         popover-trigger="hover"/>
+
     <div class="card-body">
         <h4>
             <a :href="dataset.page" :title="dataset.title">
-                {{ dataset.title | truncate 80 }}
+                {{ dataset.title }}
             </a>
         </h4>
-    </div>
-    <footer>
-        <ul>
-            <li v-if="dataset.spatial && dataset.spatial.zones && dataset.spatial.zones.length > 0">
-                <a class="btn btn-xs" v-tooltip tooltip-placement="top" :title="_('Territorial coverage')">
-                    <span class="fa fa-map-marker fa-fw"></span>
-                    {{ dataset.spatial.zones[0].name }}
-                </a>
-            </li>
-            <li v-if="dataset.metrics">
-                <a class="btn btn-xs" v-tooltip tooltip-placement="top" :title="_('Reuses')">
-                    <span class="fa fa-retweet fa-fw"></span>
-                    {{ dataset.metrics.reuses || 0 }}
-                </a>
-            </li>
-            <li v-if="dataset.metrics">
-                <a class="btn btn-xs" v-tooltip tooltip-placement="top" :title="_('Stars')">
-                    <span class="fa fa-star fa-fw"></span>
-                    {{ dataset.metrics.followers || 0 }}
-                </a>
-            </li>
-        </ul>
-    </footer>
 
-    <a class="rollover fade in" :href="dataset.page"
-        :title="dataset.title">
-        {{{ dataset.description | markdown 180 }}}
-    </a>
-    <footer class="rollover fade in">
+        <div class="clamp-3">{{{ dataset.description | markdown 180 }}}</div>
+    </div>
+
+    <footer class="card__footer">
         <ul>
+            <li v-tooltip :title="_('Resources count')">
+                <span class="fa fa-files-o fa-fw"></span>
+                {{ dataset.resources.length  }}
+            </li>
+            <li v-if="dataset.spatial && dataset.spatial.zones" v-tooltip :title="_('Territorial coverage')">
+                <span class="fa fa-map-marker fa-fw"></span>
+                {{ dataset.spatial.zones[0].name }}
+            </li>
+            <li v-if="dataset.metrics" v-tooltip :title="_('Reuses')">
+                <span class="fa fa-retweet fa-fw"></span>
+                {{ dataset.metrics.reuses || 0 }}
+            </li>
+            <li v-if="dataset.metrics" v-tooltip :title="_('Stars')">
+                <span class="fa fa-star fa-fw"></span>
+                {{ dataset.metrics.followers || 0 }}
+            </li>
+
             <!-- Temporal coverage -->
-            <li v-if="dataset.temporal_coverage">
-                <a class="btn btn-xs" v-tooltip tooltip-placement="top" :title="_('Temporal coverage')">
-                    <span class="fa fa-calendar fa-fw"></span>
-                    {{ dataset.temporal_coverage | daterange }}
-                </a>
+            <li v-if="dataset.temporal_coverage" v-tooltip :title="_('Temporal coverage')">
+                <span class="fa fa-calendar fa-fw"></span>
+                {{ dataset.temporal_coverage | daterange }}
             </li>
 
             <!-- Territorial coverage -->
-            <li v-if="dataset.spatial && dataset.spatial.granularity">
-                <a class="btn btn-xs" v-tooltip tooltip-placement="top"
-                    :title="_('Territorial coverage granularity')">
-                    <span class="fa fa-bullseye fa-fw"></span>
-                    {{ dataset | granularity_label }}
-                </a>
-            </li>
-
-            <!-- frequency -->
-            <li v-if="dataset.frequency">
-                <a class="btn btn-xs" v-tooltip tooltip-placement="top" :title="_('Frequency')">
-                    <span class="fa fa-clock-o fa-fw"></span>
-                    {{ dataset | frequency_label }}
-                </a>
+            <li v-if="dataset.spatial && dataset.spatial.granularity" v-tooltip :title="_('Territorial coverage granularity')">
+                <span class="fa fa-bullseye fa-fw"></span>
+                {{ dataset | granularity_label }}
             </li>
         </ul>