+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Geopackage as default allowed file formats [#1425](https://github.com/opendatateam/udata/pull/1425)
 - Added more entrypoints and document them. There is no more automatically enabled plugin by installation. Plugins can now properly contribute translations. [#1431](https://github.com/opendatateam/udata/pull/1431)
 - Soft breaks in markdown is rendered as line return as allowed by the [commonmark specifications](http://spec.commonmark.org/0.28/#soft-line-breaks), client-side rendering follows the same security rules [#1432](https://github.com/opendatateam/udata/pull/1432)
+- Update card components to make them more consistent [#1383](https://github.com/opendatateam/udata/pull/1383)
 
 ## 1.2.11 (2018-02-05)